:recycle: Switch from using bulkReplaceDatasetEntries model method to sub-service pattern.
