+++ api/src/services/dataset-integrations/index.ts
@@ -3,4 +3,5 @@ export { UpdateService } from "./update-service"
 
 // Sub-Services
 export { ActivateService } from "./activate-service"
+export { BulkReplaceDatasetEntriesService } from "./bulk-replace-dataset-entries-service"
 export { RefreshService } from "./refresh-service"
