+++ api/src/services/dataset-integrations/update-service.ts
@@ -1,5 +1,5 @@
 import db, { DatasetIntegration, User } from "@/models"
-import { RefreshService } from "@/services/dataset-integrations"
+import { BulkReplaceDatasetEntriesService, RefreshService } from "@/services/dataset-integrations"
 
 import BaseService from "@/services/base-service"
 
@@ -27,7 +27,7 @@ export class UpdateService extends BaseService {
       await RefreshService.perform(this.datasetIntegration)
       await this.datasetIntegration.applyJMESPathTransform()
       // TODO: create fields if none exist during dataset import
-      await this.datasetIntegration.bulkReplaceDatasetEntries()
+      await BulkReplaceDatasetEntriesService.perform(this.datasetIntegration)
 
       // TODO: log user action
